feat: Store CRUD updated
전체 조회시 영업중 상태인 가게들만 조회 가능
단건 조회시 메뉴도 같이 출력
가게 상태 boolean에서 enum타입으로 변경(총 3가지 상태로 나뉨. NOTOPENED, OPENED, SHUTDOWN)

+++ deliveryApp/src/main/java/com/example/deliveryApp/menu/dto/MenuRequestDto.java
@@ -10,9 +10,9 @@ public class MenuRequestDto {
 
     private String menuName;
     private int price;
-    private int storeId;
+    private Long storeId;
 
-    public MenuRequestDto(String menuName, int price, int storeId) {
+    public MenuRequestDto(String menuName, int price, Long storeId) {
         this.menuName = menuName;
         this.price = price;
         this.storeId = storeId;
